Switch to a monolithic file to hold parsing classes.
