+++ src/bod_diagnostics/_utils.py
